refactor: extract demo session fixtures and fix lint warnings
- Extract demo session factory from useGameplayStore.ts to src/__fixtures__/gameplay/
- Creates centralized fixture functions for demo game data
- Reduce store from 511 to 194 lines (~300 lines moved to fixtures)
- Fix 8 unused variable/import lint warnings in gameplay components
- Prefix reserved functions with underscore (_pixelToHex, _getDamagePercent)

Improves store maintainability by separating demo/test data from production code.
