+++ src/__fixtures__/gameplay/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Gameplay Fixtures Index
+ * Export barrel for gameplay-related test fixtures.
+ */
+
+export {
+  createDemoSession,
+  createDemoWeapons,
+  createDemoMaxArmor,
+  createDemoMaxStructure,
+  createDemoPilotNames,
+  createDemoHeatSinks,
+} from './demoSession';
