+++ src/components/gameplay/GameplayLayout.tsx
@@ -10,9 +10,7 @@ import {
   GamePhase,
   GameSide,
   IGameSession,
-  IGameState,
   IUnitGameState,
-  IGameEvent,
   IUnitToken,
   ILayoutConfig,
   IWeaponStatus,
